change ArticleDetail to container
