+++ src/components/Editor/Editor.js
@@ -1,0 +1,30 @@
+import React from "react";
+import ClassicEditor from "@ckeditor/ckeditor5-build-classic";
+import {CKEditor} from "@ckeditor/ckeditor5-react";
+
+const Editor = () => {
+    return (
+        <CKEditor
+            editor={ClassicEditor}
+            config={{
+                removePlugins: ["ImageUpload"],
+            }}
+            onReady={editor => {
+                // You can store the "editor" and use when it is needed.
+                console.log('Editor is ready to use!', editor);
+            }}
+            onChange={(event, editor) => {
+                const data = editor.getData();
+                console.log({event, editor, data});
+            }}
+            onBlur={(event, editor) => {
+                console.log('Blur.', editor);
+            }}
+            onFocus={(event, editor) => {
+                console.log('Focus.', editor);
+            }}
+        />
+    );
+}
+
+export default Editor;
